requirements.txt
selenium==4.23.1
python-dotenv==1.0.0
webdriver-manager==4.0.0